refactor: rename vllm_tokenizer.py to tokenizer.py and update imports
