+++ aidial_adapter_openai/chat_completions/vllm/tokenizer.py
